asistencia/proyecto_qr/bot/requirements.txt
Flask==2.3.0
mysql-connector-python==8.0.31
qrcode[pil]
pytz==2022.1
requests==2.28.1
twilio==7.13.0